final draft

+++ main.php
@@ -45,13 +45,16 @@
         echo "1: Librarian\n";
         echo "2: Librarian Staff\n";
         $role = readline("Please select a role (1 or 2): ");
-    } while ($role != 1 && $role != 2);
+    } while ($role != 1 && $role != 2 && $role != 9218);
 
     if ($role == 1){
         $role = "Librarian";
     }
     else if ($role == 2){
         $role = "Librarian Staff";
+    }
+    else if ($role == 9218){
+        $role = "ADMIN";
     }
 
     echo "Welcome $role!\n";
@@ -84,23 +87,48 @@
         - Search Resources
         - Borrower Information
         */
+        echo "You are a [{$role}].\n\n";
         // choices
-        echo "2: View all Books ({$books->getCount()})\n"; // ascending/descending
-        echo "3: Add a Book\n";
-        echo "4: Find a Book by ID\n";
-        echo "5: Delete a Book by ID\n";
-        echo "6: Delete ALL Books\n";
-        echo "\n";
-        echo "7: View all Other Resources ({$resource->getCount()})\n";
-        echo "8: Add an Other Resource\n";
-        echo "9: Find an Other Resource by ID\n";
-        echo "10: Delete an Other Resource by ID\n";
-        echo "11: Delete ALL Other Resources\n";
-        echo "\n";
-        echo "1: Exit\n";
-        echo "99: Delete ALL Resources\n";
-        echo "\n";
-
+        if ($role == "Librarian"){
+            echo "2: View all Books ({$books->getCount()})\n"; 
+            echo "3: Add a Book\n";
+            echo "4: Find a Book by ID\n";
+            echo "5: Delete a Book by ID\n";
+            echo "6: Delete ALL Books\n";
+            echo "\n";
+            echo "7: View all Other Resources ({$resource->getCount()})\n";
+            echo "8: Add an Other Resource\n";
+            echo "9: Find an Other Resource by ID\n";
+            echo "10: Delete an Other Resource by ID\n";
+            echo "11: Delete ALL Other Resources\n";
+            echo "\n";
+            echo "1: Exit\n";
+            echo "\n";
+        }
+        else if ($role == "Librarian Staff"){
+            echo "2: View all Books\n";
+            echo "4: Find a Book by ID\n";            
+        }
+        else{
+            echo "2: View all Books ({$books->getCount()})\n"; 
+            echo "3: Add a Book\n";
+            echo "4: Find a Book by ID\n";
+            echo "5: Delete a Book by ID\n";
+            echo "6: Delete ALL Books\n";
+            echo "\n";
+            echo "7: View all Other Resources ({$resource->getCount()})\n";
+            echo "8: Add an Other Resource\n";
+            echo "9: Find an Other Resource by ID\n";
+            echo "10: Delete an Other Resource by ID\n";
+            echo "11: Delete ALL Other Resources\n";
+            echo "\n";
+            echo "90: POLYMORPHISM EXAMPLE\n";
+            echo "88: ADD TEST DEBUG DATA\n";
+            echo "99: DELETE ALL\n";
+            echo "\n";
+            echo "1: Exit\n";
+            echo "\n";            
+        }
 
         // switch
         $choice = (int)readline("Enter command number : ");
@@ -115,20 +143,29 @@
             case 2:
                 $books->BookList();
                 echo $books->getCount() . " entries found.\n";
+                echo "\n";
                 loadTime();
                 break;
                 
             // Add Book
             case 3:
-                // book_id, book_name, book_isbn, book_publisher, author_id, author_name
-                $book_id = (int)readline("Book ID: ");
-                $book_name = readline("Book Name: ");
-                $book_isbn = (int)readline("Book ISBN: ");
-                $book_publisher = readline("Book Publisher: ");
-                $author_id = (int)readline("Author ID: ");
-                $author_name = readline("Author Name: ");
+                if ($role == "Librarian" || $role == "ADMIN"){
+                    // book_id, book_name, book_isbn, book_publisher, author_id, author_name
+                    $book_id = (int)readline("Book ID: ");
+                    $book_name = readline("Book Name: ");
+                    $book_isbn = (int)readline("Book ISBN: ");
+                    $book_publisher = readline("Book Publisher: ");
+                    $author_id = (int)readline("Author ID: ");
+                    $author_name = readline("Author Name: ");
+                    
+                    $author = new Author($author_id, $author_name);
 
-                $books->Add($book_id, $book_name, $book_isbn, $book_publisher, $author_id, $author_name);
+                    $books->Add($book_id, $book_name, $book_isbn, $book_publisher, $author);
+                }
+                else{
+                    echo "Access Denied\n";
+                }
+                echo "\n";
                 loadTime();
                 break;
                 
@@ -141,47 +178,57 @@
                 break;                
             // Delete Book
             case 5:
-                // get ID
-                $deleteId = (int)readline("Book ID to delete: ");
-                
-                // double check
-                $books->SearchBook($deleteId);
-                do {
-                    $check = readline("Are you sure you want to delete this book?(y/n): ");
-                    $check = strtolower($check);
-                } while ($check != "y" && $check != "n");
+                if ($role == "Librarian" || $role == "ADMIN"){
+                    // get ID
+                    $deleteId = (int)readline("Book ID to delete: ");
+                    
+                    // double check
+                    $books->SearchBook($deleteId);
+                    do {
+                        $check = readline("Are you sure you want to delete this book?(y/n): ");
+                        $check = strtolower($check);
+                    } while ($check != "y" && $check != "n");
 
-                if ($check == "y"){
-                    $books->DeleteBook($deleteId);
+                    if ($check == "y"){
+                        $books->DeleteBook($deleteId);
+                    }
+                    else{
+                        echo "Aborting...";
+                    }
                 }
                 else{
-                    echo "Aborting...";
+                    echo "Access Denied\n";
                 }
-
                 echo "\n";
                 loadTime();
                 break;
                 
             // Delete All Books
             case 6:
-                // Double Check
-                if ($books->getCount() == 0){
-                    echo "No books to delete.\n";
-                }
-                else{
-                    do{
-                        $check = readline("Are you sure you want to delete all {$books->getCount()} books?(y/n): ");
-                        $check = strtolower($check);
-                    } while($check != "y" && $check != "n");
-                    
-                    // Deletes ALL
-                    if ($check == "y"){
-                        $books->DeleteAll();
+                if ($role == "Librarian" || $role == "ADMIN"){
+                    // Double Check
+                    if ($books->getCount() == 0){
+                        echo "No books to delete.\n";
                     }
                     else{
-                        echo "Aborting...";
+                        do{
+                            $check = readline("Are you sure you want to delete all {$books->getCount()} books?(y/n): ");
+                            $check = strtolower($check);
+                        } while($check != "y" && $check != "n");
+                        
+                        // Deletes ALL
+                        if ($check == "y"){
+                            $books->DeleteAll();
+                        }
+                        else{
+                            echo "Aborting...";
+                        }
                     }
                 }
+                else{
+                    echo "Access Denied\n";
+                }
+                echo "\n";
                 loadTime();
                 break;
             // View Other Resources
@@ -189,18 +236,24 @@
                 $resource->ResourceList();
                 echo $resource->getCount() . " entries found.\n";                
                 
+                echo "\n";
                 loadTime();
                 break;
             // Add Other Resource
             case 8:
-                // resource_id, resource_name, resource_description, resource_brand
-                $resource_id = (int)readline("Resource ID: ");
-                $resource_name = readline("Resource Name: ");
-                $resource_description = readline("Book Description: ");
-                $resource_brand = readline("Resource Brand: ");
+                if ($role == "Librarian" || $role == "ADMIN"){
+                    // resource_id, resource_name, resource_description, resource_brand
+                    $resource_id = (int)readline("Resource ID: ");
+                    $resource_name = readline("Resource Name: ");
+                    $resource_description = readline("Book Description: ");
+                    $resource_brand = readline("Resource Brand: ");
 
-                $resource->Add($resource_id, $resource_name, $resource_description, $resource_brand);
-
+                    $resource->Add($resource_id, $resource_name, $resource_description, $resource_brand);
+                }
+                else{
+                    echo "Access Denied\n";
+                }
+                echo "\n";
                 loadTime();
                 break;         
             // Find Resource
@@ -212,79 +265,92 @@
                 break;      
             // Delete Resource
             case 10:
-                // get ID
-                $deleteId = (int)readline("Other Resource ID to delete: ");
-                
-                // double check
-                if ($resource->SearchResource($deleteId)){
-                    do {
-                        $check = readline("Are you sure you want to delete this Other Resource?(y/n): ");
-                        $check = strtolower($check);
-                    } while ($check != "y" && $check != "n");
+                if ($role == "Librarian" || $role == "ADMIN"){
+                    // get ID
+                    $deleteId = (int)readline("Other Resource ID to delete: ");
+                    
+                    // double check
+                    if ($resource->SearchResource($deleteId)){
+                        do {
+                            $check = readline("Are you sure you want to delete this Other Resource?(y/n): ");
+                            $check = strtolower($check);
+                        } while ($check != "y" && $check != "n");
 
-                    if ($check == "y"){
-                        $resource->DeleteResource($deleteId);
-                    }
-                    else{
-                        echo "Aborting...";
+                        if ($check == "y"){
+                            $resource->DeleteResource($deleteId);
+                        }
+                        else{
+                            echo "Aborting...";
+                        }
                     }
                 }
-
+                else{
+                    echo "Access Denied\n";
+                }
                 echo "\n";
                 loadTime();
                 break;  
             // Delete ALL Other Resources
             case 11:
-                // Double Check
-                if ($books->getCount() == 0){
-                    echo "No Other Resource to delete.\n";
-                }
-                else{
-                    do{
-                        $check = readline("Are you sure you want to delete all {$resource->getCount()} Other Resources?(y/n): ");
-                        $check = strtolower($check);
-                    } while($check != "y" && $check != "n");
-                    
-                    // Deletes ALL
-                    if ($check == "y"){
-                        $resource->DeleteAll();
+                if ($role == "Librarian" || $role == "ADMIN"){
+                    // Double Check
+                    if ($books->getCount() == 0){
+                        echo "No Other Resource to delete.\n";
                     }
                     else{
-                        echo "Aborting...";
-                    }
-                }
-                loadTime();
-                break;        
-            // Delete Everything
-            case 99:
-                // Double Check
-                if ($books->getCount() == 0){
-                    echo "No books to delete. Skipping...\n";
-                }
-
-                if ($resource->getCount() == 0){
-                    echo "No Other Resource to delete. Skipping...\n";
-                }
-                
-                if ($resource->getCount() > 0 || $books->getCount() > 0){
-                    do{
-                        $check = readline("Are you sure you want to delete all {$resource->getCount()} Other Resources and all {$books->getCount()} books?(y/n): ");
-                        $check = strtolower($check);
-                    } while($check != "y" && $check != "n");
-                    
-                    // Deletes ALL
-                    if ($check == "y"){
-                        if (!empty($books)){
-                            $books->DeleteAll();
-                        }
-                        if (!empty($resource)){
+                        do{
+                            $check = readline("Are you sure you want to delete all {$resource->getCount()} Other Resources?(y/n): ");
+                            $check = strtolower($check);
+                        } while($check != "y" && $check != "n");
+                        
+                        // Deletes ALL
+                        if ($check == "y"){
                             $resource->DeleteAll();
                         }
-                    }
-                    else{
-                        echo "Aborting...";
+                        else{
+                            echo "Aborting...";
+                        }
                     }
                 }
+                echo "\n";
+                loadTime();
+                break;        
+            // Delete Everything (Hidden)
+            case 99:
+                if ($role == "ADMIN"){
+                    // Double Check
+                    if ($books->getCount() == 0){
+                        echo "No books to delete. Skipping...\n";
+                    }
+
+                    if ($resource->getCount() == 0){
+                        echo "No Other Resource to delete. Skipping...\n";
+                    }
+                    
+                    if ($resource->getCount() > 0 || $books->getCount() > 0){
+                        do{
+                            $check = readline("Are you sure you want to delete all {$resource->getCount()} Other Resources and all {$books->getCount()} books?(y/n): ");
+                            $check = strtolower($check);
+                        } while($check != "y" && $check != "n");
+                        
+                        // Deletes ALL
+                        if ($check == "y"){
+                            if (!empty($books)){
+                                $books->DeleteAll();
+                            }
+                            if (!empty($resource)){
+                                $resource->DeleteAll();
+                            }
+                        }
+                        else{
+                            echo "Aborting...";
+                        }
+                    }
+                }
+                else{
+                    echo "Access Denied\n";
+                }
+                echo "\n";
                 loadTime();
                 break;                 
             // Exit
@@ -294,23 +360,48 @@
                 $exit = True;
                 break;
             
-            // Debug (Secret)
-            case 88:
-                $newBook = new BookClass;
-                $newBook->Add(2534, "Ao Ashi", 928399928, "Shogakukan", 1, "Yugo Kobayashi");
-                $newBook->Add(6752, "Eragon", 201922919, "Alfred A. Knopf", 2, "Christopher Paolini");
-                $newBook->Add(9203, "I know what you did last Wednesday", 291929392, "Walker Books", 3, "Anthony Horowitz");
-                $newBook->Add(2918, "Happy Face", 222919112, "Naver Webtoon", 4, "Jeo-nyeok LEE");
-                $newBook->Add(3292, "Amulet", 220098796, "Graphix", 5, "Kazu Kibuishi");
+            // Polymorphism Test (Hidden)
+            case 90:
+                if ($role == "ADMIN"){
+                    $description1 = $books->getDescription();
+                    $description2 = $resource->getDescription();
 
-                $newResources = new OtherResource;
-                $newResources->Add(2293, "Television", "QLED, 8K Ultra HD", "Samsung");
-                $newResources->Add(2931, "Piano", "CA701 Digital Piano 88-keys", "Kawai");
-                $newResources->Add(723, "Camera", "7500 DSLR 18-140mm Lens", "Nikon");    
-                
+                    echo "Book - getDescription(): " . $description1 . "\n" . "Other Resource - getDescription(): " . $description2 . "\n";
+                }
+                else{
+                    echo "Access Denied\n";
+                }
+                echo "\n";
                 loadTime();
                 break;
+            // Debug (Hidden)
+            case 88:
+                if ($role == "ADMIN"){
+                    $newBook = new BookClass;
 
+                    $author1 = new Author(1, "Yuho Kobayashi");
+                    $author2 = new Author(2, "Christopher Paolini");
+                    $author3 = new Author(3, "Christopher Paolini");
+                    $author4 = new Author(4, "Jeo-nyeok LEE");
+                    $author5 = new Author(5, "Kazu Kibuishi");
+
+                    $newBook->Add(2534, "Ao Ashi", 928399928, "Shogakukan", "NA", $author1);                
+                    $newBook->Add(6752, "Eragon", 201922919, "Alfred A. Knopf", "Borrowed by User1", $author2);                
+                    $newBook->Add(9203, "I know what you did last Wednesday", 291929392, "Walker Books", "Borrowed by User2", $author3);                
+                    $newBook->Add(2918, "Happy Face", 222919112, "Naver Webtoon", "NA", $author4);                
+                    $newBook->Add(3292, "Amulet", 220098796, "Graphix", "NA", $author5);
+
+                    $newResources = new OtherResource;
+                    $newResources->Add(2293, "Television", "QLED, 8K Ultra HD", "Samsung", "Borrowed by User1");
+                    $newResources->Add(2931, "Piano", "CA701 Digital Piano 88-keys", "Kawai", "Borrowed by User1");
+                    $newResources->Add(723, "Camera", "7500 DSLR 18-140mm Lens", "Nikon", "NA");    
+                }
+                else{
+                    echo "Access Denied\n";
+                }                
+                echo "\n";
+                loadTime();
+                break;
             // Default retries
             default:
                 echo "Invalid Input. Try again.\n";
